split packges
