+++ packages/cli/README.md
@@ -1,0 +1,3 @@
+# @tokenami/cli
+
+View docs [here](https://github.com/tokenami/tokenami).
