+++ packages/cli/bin.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import './dist/bin.js';
