+++ packages/cli/package.json
@@ -1,9 +1,8 @@
 {
-  "name": "tokenami",
+  "name": "@tokenami/cli",
   "version": "0.0.95",
   "license": "MIT",
   "type": "module",
-  "main": "./dist/index.cjs",
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "repository": "tokenami/tokenami",
@@ -15,11 +14,10 @@
       "import": {
         "types": "./dist/index.d.ts",
         "default": "./dist/index.js"
-      },
-      "require": {
-        "types": "./dist/index.d.cts",
-        "default": "./dist/index.cjs"
       }
+    },
+    "./bin": {
+      "import": "./dist/bin.js"
     }
   },
   "bin": {
@@ -35,21 +33,22 @@
   "scripts": {
     "build": "tsup",
     "dev": "tsup --watch",
-    "test": "vitest --run",
-    "test:watch": "vitest",
     "typecheck": "tsc --noEmit"
   },
-  "devDependencies": {
-    "@types/node": "^20.3.1",
-    "tsup": "^8.4.0",
-    "typescript": "^5.1.3",
-    "vitest": "^0.34.6"
-  },
   "dependencies": {
-    "@tokenami/cli": "workspace:*",
-    "@tokenami/node": "workspace:*"
+    "@tokenami/config": "workspace:*",
+    "@tokenami/node": "workspace:*",
+    "browserslist": "^4.26.3",
+    "cac": "^6.7.14",
+    "chokidar": "^3.5.3",
+    "inquirer": "^9.2.12",
+    "lightningcss": "1.29.1",
+    "pathe": "^1.1.1",
+    "typescript": "^5.1.3"
   },
-  "peerDependencies": {
-    "typescript": ">= 5"
+  "devDependencies": {
+    "@types/inquirer": "^9.0.7",
+    "@types/node": "^20.3.1",
+    "tsup": "^8.4.0"
   }
 }
