+++ packages/cli/tsconfig.json
@@ -1,15 +1,14 @@
 {
   "extends": "../../tsconfig.json",
   "include": ["src"],
-  "exclude": ["src/**/*.test.ts"],
   "compilerOptions": {
     "baseUrl": ".",
     "tsBuildInfoFile": "node_modules/.cache/tsbuildinfo.json",
     "outDir": "dist",
     "target": "ES2022",
     "paths": {
-      "@tokenami/node": ["../node/src"],
-      "@tokenami/css": ["../css/src"] // used in stubs
+      "@tokenami/config": ["../config/src"],
+      "@tokenami/node": ["../node/src"]
     }
   }
 }
