+++ packages/cli/tsup.config.ts
@@ -3,11 +3,20 @@ import { defineConfig } from 'tsup';
 export default defineConfig([
   {
     entry: ['./src/index.ts'],
-    format: ['cjs', 'esm'],
+    format: ['esm'],
     splitting: false,
     treeshake: true,
     clean: true,
     shims: true,
     dts: true,
+  },
+  {
+    entry: ['./src/bin.ts'],
+    format: ['esm'],
+    splitting: false,
+    treeshake: true,
+    clean: false,
+    shims: true,
+    dts: false,
   },
 ]);
