+++ packages/css/tsconfig.json
@@ -8,6 +8,7 @@
     "target": "ES2022",
     "paths": {
       "@tokenami/config": ["../config/src"],
+      "@tokenami/node": ["../node/src"],
       "tokenami": ["../tokenami/src"]
     }
   }
