+++ packages/node/README.md
@@ -1,0 +1,3 @@
+# @tokenami/node
+
+View docs [here](https://github.com/tokenami/tokenami).
