+++ packages/node/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "tokenami",
+  "name": "@tokenami/node",
   "version": "0.0.95",
   "license": "MIT",
   "type": "module",
@@ -7,6 +7,7 @@
   "module": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "repository": "tokenami/tokenami",
+  "sideEffects": false,
   "publishConfig": {
     "access": "public"
   },
@@ -22,15 +23,9 @@
       }
     }
   },
-  "bin": {
-    "tokenami": "bin.js"
-  },
-  "engines": {
-    "node": ">=20"
-  },
   "files": [
     "dist",
-    "bin.js"
+    "stubs"
   ],
   "scripts": {
     "build": "tsup",
@@ -39,15 +34,27 @@
     "test:watch": "vitest",
     "typecheck": "tsc --noEmit"
   },
+  "dependencies": {
+    "@stitches/stringify": "^1.2.8",
+    "@tokenami/config": "workspace:*",
+    "acorn": "^8.11.3",
+    "acorn-walk": "^8.3.2",
+    "css-tree": "^3.0.0",
+    "csstype": "^3.1.2",
+    "culori": "^4.0.1",
+    "fast-glob": "^3.2.12",
+    "jiti": "^1.21.0",
+    "lightningcss": "1.29.1",
+    "pathe": "^1.1.1",
+    "sucrase": "^3.34.0"
+  },
   "devDependencies": {
+    "@types/css-tree": "^2.3.8",
+    "@types/culori": "^2.1.0",
     "@types/node": "^20.3.1",
     "tsup": "^8.4.0",
     "typescript": "^5.1.3",
     "vitest": "^0.34.6"
-  },
-  "dependencies": {
-    "@tokenami/cli": "workspace:*",
-    "@tokenami/node": "workspace:*"
   },
   "peerDependencies": {
     "typescript": ">= 5"
