+++ packages/node/src/declarations.ts
