+++ packages/node/src/diagnostics.test.ts
