+++ packages/node/src/ts-plugin/completions.ts
