+++ packages/node/stubs/tokenami.config.cjs
