+++ packages/tokenami/bin.js
@@ -1,2 +1,2 @@
 #!/usr/bin/env node
-import './dist/cli.js';
+import '@tokenami/cli/bin';
